API for dish products and reports fixed

+++ backend-springboot/src/main/java/com/wellbeeing/wellbeeing/api/diet/DishController.java
@@ -1,8 +1,10 @@
 package com.wellbeeing.wellbeeing.api.diet;
 
 import com.wellbeeing.wellbeeing.domain.diet.Dish;
+import com.wellbeeing.wellbeeing.domain.exception.ForbiddenException;
 import com.wellbeeing.wellbeeing.domain.exception.NotFoundException;
 import com.wellbeeing.wellbeeing.domain.message.PaginatedResponse;
+import com.wellbeeing.wellbeeing.service.account.UserService;
 import com.wellbeeing.wellbeeing.service.diet.DishService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -12,6 +14,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.lang.NonNull;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.UUID;
 
@@ -19,10 +22,13 @@ import java.util.UUID;
 @RestController
 public class DishController {
     public DishService dishService;
+    public UserService userService;
 
     @Autowired
-    public DishController(@Qualifier("dishService") DishService dishService){
+    public DishController(@Qualifier("dishService") DishService dishService,
+                          @Qualifier("userService") UserService userService){
         this.dishService = dishService;
+        this.userService = userService;
     }
 
     @RequestMapping(path = "/dish/{dishId}", method = RequestMethod.GET)
@@ -31,6 +37,7 @@ public class DishController {
         return new ResponseEntity<>(actDish, HttpStatus.OK);
     }
 
+    //@RolesAllowed(ERole.Name.ROLE_ADMIN)
     @RequestMapping(path = "/dish/{dishId}/update-derived", method = RequestMethod.GET)
     public ResponseEntity<?> updateDishDerivedById(@PathVariable("dishId") UUID dishID) throws NotFoundException {
         boolean resp = dishService.updateCaloriesAndMacrosByDishId(dishID);
@@ -53,7 +60,7 @@ public class DishController {
     @RequestMapping(path = "/dish/search", method = RequestMethod.GET)
     public ResponseEntity<?> getDishesWithNameLike(@RequestParam(value = "page", defaultValue = "0") int page,
                                                      @RequestParam(value = "size", defaultValue = "10") int size,
-                                                     @RequestParam(value = "nameLike", defaultValue = "a") String nameLike){
+                                                     @RequestParam(value = "nameLike", defaultValue = "") String nameLike){
         Page<Dish> dishesPage = dishService.getDishesWithNameLike(nameLike, size, page);
         PaginatedResponse response = PaginatedResponse.builder()
                     .currentPage(dishesPage.getNumber())
@@ -64,18 +71,36 @@ public class DishController {
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    @RequestMapping(path = "/dish", method = RequestMethod.POST)
-    public ResponseEntity<?> addDish(@RequestBody @NonNull Dish dish)  {
-        return new ResponseEntity<>(dishService.addDish(dish), HttpStatus.OK);
+    //@RolesAllowed(ERole.Name.ROLE_DIETICIAN)
+    @RequestMapping(path = "/dish/dietician", method = RequestMethod.POST)
+    public ResponseEntity<?> addDish(Principal principal, @RequestBody @NonNull Dish dish) throws NotFoundException {
+        return new ResponseEntity<>(dishService.addDish(dish, userService.findUserIdByUsername(principal.getName())), HttpStatus.OK);
     }
 
-    @RequestMapping(path = "/dish/{dishId}", method = RequestMethod.PUT)
-    public ResponseEntity<?> updateDish(@RequestBody @NonNull Dish dish, @PathVariable("dishId") UUID dishId)  {
+    //@RolesAllowed(ERole.Name.ROLE_DIETICIAN)
+    @RequestMapping(path = "/dish/{dishId}/dietician", method = RequestMethod.PUT)
+    public ResponseEntity<?> updateDish(Principal principal, @RequestBody @NonNull Dish dish, @PathVariable("dishId") UUID dishId) throws NotFoundException, ForbiddenException {
+        UUID dieticianId = userService.findUserIdByUsername(principal.getName());
+        Dish actDish = dishService.getDishById(dishId);
+        if(!actDish.getDishCreator().getId().equals(dieticianId))
+            throw new ForbiddenException("Access to dish with id: " + dishId + "forbidden");
         return new ResponseEntity<>(dishService.updateDish(dish, dishId), HttpStatus.OK);
     }
 
-    @RequestMapping(path = "/dish/{dishId}", method = RequestMethod.DELETE)
-    public ResponseEntity<?> addProduct(@PathVariable("dishId") UUID dishId)  {
+    //@RolesAllowed(ERole.Name.ROLE_DIETICIAN)
+    @RequestMapping(path = "/dish/dietician", method = RequestMethod.GET)
+    public ResponseEntity<?> getDieticianDishes(Principal principal){
+        UUID dieticianId = userService.findUserIdByUsername(principal.getName());
+        return new ResponseEntity<>(dishService.getDieticianDishesByDieticianId(dieticianId), HttpStatus.OK);
+    }
+
+    //@RolesAllowed(ERole.Name.ROLE_DIETICIAN)
+    @RequestMapping(path = "/dish/{dishId}/dietician", method = RequestMethod.DELETE)
+    public ResponseEntity<?> deleteDish(Principal principal, @PathVariable("dishId") UUID dishId) throws ForbiddenException, NotFoundException {
+        UUID dieticianId = userService.findUserIdByUsername(principal.getName());
+        Dish actDish = dishService.getDishById(dishId);
+        if(!actDish.getDishCreator().getId().equals(dieticianId))
+            throw new ForbiddenException("Access to dish with id: " + dishId + "forbidden");
         return new ResponseEntity<>(dishService.deleteDish(dishId), HttpStatus.OK);
     }
